pyproject.toml
[project]
name = "dspy-reference-examples"
version = "0.1.0"
description = "DSPy optimization reference for Ozempic complaint classification"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "dspy-ai>=3.1.0",
  "openai>=2.14.0",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.12.0",
  "fastapi>=0.128.0",
  "uvicorn>=0.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
  "ruff>=0.14.10",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
  "integration: tests that hit real LLMs or external services",
]
log_cli = true
log_cli_level = "INFO"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
]
ignore = [
  "E501",  # Line too long - often not auto-fixable, especially in strings
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
